pyproject.toml
[project]
name = "omnimodal-agent-memory"
version = "0.1.0"
description = "A minimal reference implementation of a shared multimodal memory layer for AI agents using Gemini Embeddings."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.67.0",
    "qdrant-client>=1.17.1",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.41.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "numpy>=2.4.3",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.6",
    "black>=26.3.1",
    "pytest>=9.0.2",
    "httpx>=0.28.1",
]
samples = [
    "Pillow>=12.1.1",
    "fpdf2>=2.8.7",
    "gtts>=2.5.4",
    "imageio[ffmpeg]>=2.37.3",
]
notebook = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "scikit-learn>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
target-version = ["py312"]
line-length = 88
